requirements.txt
flet
requests
supabase
python-dotenv
